% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-rules.R
\name{apply_rules}
\alias{apply_rules}
\title{Apply rules}
\usage{
apply_rules(word, name, steprules)
}
\arguments{
\item{word}{word to which you want to apply the rules}

\item{name}{the rule name, possible values are: 'Plural', 'Feminine', 'Adverb',
'Augmentative', 'Noun', 'Verb', 'Vowel' .}

\item{steprules}{steprules as obtained from the function extract_rules.}
}
\description{
Apply rules
}
