% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rules.R
\name{extract_raw_rules}
\alias{extract_raw_rules}
\title{Extract raw rules}
\usage{
extract_raw_rules(raw_rules)
}
\arguments{
\item{raw_rules}{a charcter with the raw rules.}
}
\description{
Separate the seven kinds of rules
}
