% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rules.R
\name{extract_replacement_rules}
\alias{extract_replacement_rules}
\title{Extract replacement rules}
\usage{
extract_replacement_rules(raw_repl)
}
\arguments{
\item{raw_repl}{the part with replacement rules for each step rule.}
}
\description{
Parses the the raw replacement rules.
}

