% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rules.R
\name{extract_rule_info}
\alias{extract_rule_info}
\title{Extract Rule Info}
\usage{
extract_rule_info(rule)
}
\arguments{
\item{rule}{the rule you want to extract infos}
}
\description{
Extract all info for one rule
}
