% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rules.R
\name{extract_rules}
\alias{extract_rules}
\title{Extract Rules from file}
\usage{
extract_rules(path = system.file("steprules.txt", package = "rslp"))
}
\arguments{
\item{path}{path to the raw steprules. Most of the times you don't have to change it.}
}
\description{
This function parse the rules that are disponible in the RLSP package
disponible in the this \url{http://www.inf.ufrgs.br/~arcoelho/rslp/integrando_rslp.html}
This file has been downloaded and is installed with the package. It's path
can be found using system.file("steprules.txt", package = "rslp")
A parsed version is saved is also installed with the package and its path
can be found using system.file("steprules.rds", package = "rslp").
}

