% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rules.R
\name{extract_rules_info}
\alias{extract_rules_info}
\title{Extract Rules Info}
\usage{
extract_rules_info(rules)
}
\arguments{
\item{rules}{rules parsed before by \link{extract_rule_info}}
}
\description{
Extract all info from all rules
}

