% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rslp.R
\name{rslp_}
\alias{rslp_}
\title{RSLP_}
\usage{
rslp_(word, steprules = readRDS(system.file("steprules.rds", package =
  "rslp")))
}
\arguments{
\item{word}{word to be stemmed.}

\item{steprules}{as obtained from the function extract_rules.}
}
\description{
Apply the Stemming Algorithm for the Portuguese Language to a word.
}

