% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rslp.R
\name{rslp_doc}
\alias{rslp_doc}
\title{RSLP Document}
\usage{
rslp_doc(docs, steprules = readRDS(system.file("steprules.rds", package =
  "rslp")))
}
\arguments{
\item{docs}{chr vector of documents}

\item{steprules}{as obtained from the function extract_rules. (only define if you are certain about it).
The default is to get the parsed versionof the rules installed with the package.}
}
\description{
Apply the Stemming Algorithm for the Portuguese Language to
vector of documents. It extracts words using the regex
"\\b[:alpha:]\\b"
}
\examples{
docs <- c("coma frutas pois elas fazem bem para a saúde.",
"não coma doces, eles fazem mal para os dentes.")
rslp_doc(docs)

}
\references{
\href{http://homes.dcc.ufba.br/~dclaro/download/mate04/Artigo Erick.pdf}{Article by Viviane Moreira Orengo and Christian Huyck}
}

