% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-rules.R
\name{verify_sufix}
\alias{verify_sufix}
\title{Verify}
\usage{
verify_sufix(word, rep_rules)
}
\arguments{
\item{word}{word you which to verify replacement rules}

\item{rep_rules}{data.frame of rules as specified in steprules$replacement_rule}
}
\description{
Given a list of suffixes, returns a vector of true or
false indicating if the word has each one of the suffixes.
}
